Bug fix.
Description:
Removed the overall -e because this aborted the snapshot release from happening
since it is normal not to find any .lock files outstanding when it is time
to do snapshot release.  The ls command there exits with non-zero and the
-e option aborts the whole thing unnecessary.

The -e is needed especially for the LOGDIR cleaning to prevent it from
cleaning the wrong directory (such as $HOME) if $LOGDIR is ill defined
or not existing.  Move all the cleaning code into a sub-shell to be run
with -e.  That provides the protection locally without affecting the
whole test script.